Add merge friend delete with friend get and post

+++ lynk_up_server/lynk_up_server/views.py
@@ -113,7 +113,7 @@ def event_detail(request, event_id):
     return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-@api_view(['POST', 'GET'])
+@api_view(['POST', 'GET', 'DELETE'])
 def add_friend(request, user_id):
   if request.method == 'GET':
     user = User.objects.get(id=user_id)
@@ -139,20 +139,20 @@ def add_friend(request, user_id):
     return Response(
       {"data": {"friends":serializer.data}}, status=201, content_type='application/json'
     )
+  
+  elif request.method == 'DELETE':
+    try:
+      user = User.objects.get(id=user_id)
+      friend_id = int(request.data['friend_id'])
 
-@require_http_methods(['DELETE'])
-def delete_friend(request, user_id, friend_id):
-  # import ipdb; ipdb.set_trace()
-  try:
-    user = User.objects.get(id=user_id)
-    friend = User.objects.get(id=friend_id)
-    friendship = Friend.objects.filter(user=user, friend=friend).first()
+      friend = User.objects.get(id=friend_id)
+      friendship = Friend.objects.filter(user=user, friend=friend).first()
 
-    if friendship:
-      friendship.delete()
-      return Response(status=status.HTTP_204_NO_CONTENT)
-    else:
+      if friendship:
+        friendship.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+      else:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+      
+    except User.DoesNotExist:
       return Response(status=status.HTTP_404_NOT_FOUND)
-    
-  except User.DoesNotExist:
-    return Response(status=status.HTTP_404_NOT_FOUND)
